IlmImf: Fix misc compiler warnings.
This allows IlmImf to compile cleanly with clang flags
-Wall -Wextra -Werror -Wno-unused-parameter

Fixes:
ImfOpaqueAttribute.h: 'const' type qualifier on return type has no effect [-Werror,-Wignored-qualifiers]
const int                   dataSize() const { return _dataSize; }
ImfDwaCompressor.cpp: error: comparison of integers of different signs: 'long' and 'size_t' (aka 'unsigned long') [-Werror,-Wsign-compare]
 if (cd->planarUncBufferEnd + dstScanlineSize - _planarUncBuffer[UNKNOWN] > _planarUncBufferSize[UNKNOWN] )
ImfMisc.cpp: warning: unused variable 'maxBytesPerLine' [-Wunused-variable]
    size_t maxBytesPerLine = bytesPerLineTable (header,
ImfDeepTiledInputFile.cpp:: error: comparison of integers of different signs: 'Imath_2_5::Int64' (aka 'unsigned long') and 'int' [-Werror,-Wsign-compare]
        if(_tileBuffer->dataSize != sizeOfTile)

+++ OpenEXR/IlmImf/ImfDeepTiledInputFile.cpp
@@ -605,7 +605,7 @@ TileBufferTask::execute ()
 	// sanity check data size: the uncompressed data should be exactly 
 	// 'sizeOfTile' (if it's less, the file is corrupt and there'll be a buffer overrun)
 	//
-        if(_tileBuffer->dataSize != sizeOfTile)
+        if(_tileBuffer->dataSize != Int64(sizeOfTile))
 	{
 		THROW (IEX_NAMESPACE::InputExc, "size mismatch when reading deep tile: expected " << sizeOfTile << "bytes of uncompressed data but got " << _tileBuffer->dataSize);
 	}
